dentalapp-backend/dentalapp/requirements.txt
-i https://pypi.org/simple
--extra-index-url https://pypi.org/simple/
asgiref==3.2.10
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.1
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.2
markupsafe==2.1.1
packaging==21.3
phonenumbers==8.12.11
psycopg2==2.9.3
pycparser==2.20
pyparsing==3.0.9
pytz==2020.1
pyyaml==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.15.0
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.11
